Separate implementations of RelauncherMain and RelauncherSynchronizeWithParent
